feat: add props into TrapBlock.tsx
